LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/xhr/late-upload-events-expected.txt

Adding upload event listeners after send()


FAIL Late listeners: No preflight assert_equals: expected 200 but got 0
FAIL Late listeners: Preflight assert_equals: expected 200 but got 0
